fix few compile errors

+++ explainability-service/src/main/java/org/kie/trustyai/service/payloads/metrics/drift/fouriermmd/FourierMMDMetricRequest.java
@@ -11,7 +11,6 @@ public class FourierMMDMetricRequest extends DriftMetricRequest {
     private FourierMMDParameters parameters = new FourierMMDParameters();
     private double gamma = 2.0;
 
-
     public FourierMMDParameters getParameters() {
         return parameters;
     }
@@ -20,7 +19,7 @@ public class FourierMMDMetricRequest extends DriftMetricRequest {
         this.parameters = parameters;
     }
 
-    public int getGamma() {
+    public double getGamma() {
         return gamma;
     }
